% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.svrpath.R
\name{plot.svrpath}
\alias{plot.svrpath}
\title{plot the svrpath, solution paths of SVR as a function of lambda}
\usage{
\method{plot}{svrpath}(x, intercept = FALSE, ...)
}
\arguments{
\item{x}{The svrpath object}

\item{intercept}{If it is \code{TRUE}, then an intercept path plot is given.}

\item{...}{Generic compatibility}
}
\value{
The entire solution path of SVR solution as a function of \code{lambda}.
}
\description{
produces a plot of the SVR \code{lambda} path.
}
\examples{
\donttest{
# The 'obj' is given by examples description of svrpath().
plot(obj, lty = 2, lwd = 2, col = 2, cex.lab = 1.5) }
}
\author{
Do Hyun Kim, Seung Jun Shin
}
