% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.epspath.R
\name{predict.epspath}
\alias{predict.epspath}
\title{Make predictions from an "epspath" object}
\usage{
\method{predict}{epspath}(object, newx, svr.eps = 1, ...)
}
\arguments{
\item{object}{The epspath object}

\item{newx}{Values of x to be predicted. This is a matrix with observations per row. Default is x in the epspath object.}

\item{svr.eps}{The value of the "epsilon-insensitive loss" paramter, epsilon.}

\item{...}{Generic compatibility}
}
\value{
In each case, the desired prediction.
}
\description{
Provides a prediction value at a given \code{epsilon} from \code{epspath} object.
}
\examples{
\donttest{
# The 'eobj' is given by examples description of epspath().
predict(eobj, svr.eps = .1) }
}
\author{
Do Hyun Kim, Seung Jun Shin
}
